fix. [한채선] member profile 삭제 후 readus member entity로 회원 데이터 통합 [READUS-10]
